+++ src/App.tsx
@@ -13,7 +13,7 @@ const queryClient = new QueryClient({
     },
 });
 
-import Router from './Router';
+import Router from './router';
 
 const App = () => {
     return (
